COMMERCE-6528 batch-engine-service - move to task package
